Brazil_Module/pyproject.toml
[project]
name = "Brazil_Module"
version = "1.0.0"
description = "Brazil ERP - Fiscal documents (NF-e, CT-e, NFS-e) and Banking integration for ERPNext"
authors = [
    {name = "Intelligence8", email = "contact@intelligence8.com"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    "cryptography>=46.0.0",
    "requests>=2.32.0",
    "pypdf>=4.0.0",
    "qrcode[pil]>=7.4.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["flit_core>=3.4"]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
testpaths = ["Brazil_Module/tests"]
pythonpath = ["."]

[tool.bench.frappe-dependencies]
frappe = ">=15.0.0"
erpnext = ">=15.0.0"
